Calculator - level2
공통로직 처리

+++ Sparta_PersonalAssignment/src/day1_calculator/AddOperator.java
@@ -2,7 +2,7 @@ package day1_calculator;
 
 public class AddOperator implements Operator {
     @Override
-    public int operator(int firstNumber, int secondNumber) {
+    public double operator(double firstNumber, double secondNumber) {
         return firstNumber + secondNumber;
     }
