+++ Sparta_PersonalAssignment/src/day1_calculator/DivideOperator.java
@@ -2,12 +2,13 @@ package day1_calculator;
 
 public class DivideOperator implements Operator {
     @Override
-    public int operator(int firstNumber, int secondNumber) {
+    public double operator(double firstNumber, double secondNumber) {
         if (secondNumber == 0) {
             throw new ArithmeticException("0으로 나눌 수 없습니다.");
         }
         return firstNumber / secondNumber;
     }
+
     @Override
     public String getSymbol() {
         return "/";
