+++ Sparta_PersonalAssignment/src/day1_calculator/ModOperator.java
@@ -2,7 +2,7 @@ package day1_calculator;
 
 public class ModOperator implements Operator {
     @Override
-    public int operator(int firstNumber, int secondNumber) {
+    public double operator(double firstNumber, double secondNumber) {
         if (secondNumber == 0) {
             throw new ArithmeticException("0으로 나눌 수 없습니다.");
         }
