+++ Sparta_PersonalAssignment/src/day1_calculator/MultiplyOperator.java
@@ -2,7 +2,7 @@ package day1_calculator;
 
 public class MultiplyOperator implements Operator {
     @Override
-    public int operator(int firstNumber, int secondNumber) {
+    public double operator(double firstNumber, double secondNumber) {
         return firstNumber * secondNumber;
     }
