+++ Sparta_PersonalAssignment/src/day1_calculator/Operator.java
@@ -1,6 +1,6 @@
 package day1_calculator;
 
 public interface Operator {
-    int operator(int firstNumber, int secondNumber);
+    double operator(double firstNumber, double secondNumber);
     String getSymbol();  //연산자
 }
