+++ Sparta_PersonalAssignment/src/day1_calculator/SubtractOperator.java
@@ -2,9 +2,10 @@ package day1_calculator;
 
 public class SubtractOperator implements Operator {
     @Override
-    public int operator(int firstNumber, int secondNumber) {
+    public double operator(double firstNumber, double secondNumber) {
         return firstNumber - secondNumber;
     }
+
     @Override
     public String getSymbol() {
         return "-";
